version 0004: 把所有的成績都當作 float 來處理，並修改相關的資料型態。 輸出的總分與平均皆調整為顯示到小數點後第二位

+++ Student.py
@@ -1,5 +1,5 @@
 class Student():
-    def __init__(self, name: str, grade1: int, grade2: int, grade3: int):
+    def __init__(self, name: str, grade1: float, grade2: float, grade3: float):
         self.name = name
         self.grade1 = grade1
         self.grade2 = grade2
