+++ Student_new.py
@@ -3,7 +3,7 @@ class Student():
         self.first_name = first_name
         self.second_name = second_name
         self.telephone_number = telephone_number
-        self.grade_list: list[list[int]] = []
+        self.grade_list: list[list[float]] = []
 
-    def add_grade(self, grade: list[int]) -> None:
+    def add_grade(self, grade: list[float]) -> None:
         self.grade_list.append(grade)
